requirements.txt
rfc3339
gunicorn
Django
django-allauth
django-b2
django-debug-toolbar
django-debug-toolbar-template-profiler
django-polymorphic
django-money == 2.1.1
django-rest-framework
django-react-templatetags
django-cors-headers
django-storages[azure]

azure-common == 1.1.28
boto3

django-imagekit

python-dotenv
split-settings

psycopg2-binary

squareup
wagtail
whitenoise

django-treewidget
django-mathfilters


stripe
django-address
django-mptt

pandas

django-jquery

checkdigit

django-extensions
pyparsing
pydot
py-moneyed
djangorestframework
pilkit
Pillow
requests
xlrd
beautifulsoup4
django-treebeard
django-oscar==3.1
pytz

python-dateutil
setuptools
six
b2sdk
